Specify the decidable_eq to Use When Applying finset.mem_insert_self
Since at the point pick_one_lo is defined Lean only knows there is a
decidable equality derived from the linear order on α, that decidable
equality is used in the insert coming from the pick_one_lo lemma. In
order to match the decidable equalities at the time
finset.mem_insert_self is used, one needs to specify all the implicit
parameters.

It should be possible to finish the proof from here.