Introducing a General Version of Ramsey's Property
This is essentially the definition of the multicolor Ramsey
property. In order to keep things sane in this file, I also introduced
Ramsey₂_prop and renamed most things that are speficid to 2-color
Ramsey numbers using Ramsey₂_*.

Most proofs are fixed, and from the broken proofs most are broken in
trivial ways, with the exception of fiendship_upper_bound and
Ramsey₂_prop_ineq.