Introducing monochromatic_vicinity
The definition is the neighborhood of v that is connected through
edges of a specified color. We also prove a Ramsey property of a
monochromatic vicinity of large-enough size which is the heart of the
proof of Ramsey₂_prop_ineq. The proof of Ramsey₂_prop_ineq does not
become much simpler so the payoff to have this additional definition
may come when we prove multicolor versions of that theorem. In the
immediate future, this definition will be used to prove the strict
Ramsey inequality.

Perhaps we should add this definition to the simp set?