Completed a Proof of vdW 2 3 = 9
The goal was to prove the theorem by hook or by crook in order to at
least establish the feasibility of provign van der Waerden numbers in
Lean 4. A lot of the automation can be made more elegant but with this
proof we at least know what needs to be done in order to achieve a
reasonable proof.

Given the order in which these lemmas were proven and how involved the
proofs were it is not unlikely that vdwByList could be greatly
simplified by using the closed form obtained in isArithProgIffGet.