Finished Proof for monochromaticVicinity_Ramsey
The proof indeed required us to get rid of Quotient.out. The trick
here was to define the mapping function in terms of Sym2.map and then
use some heavy notation to make everything fit. Surprisingly (or
perhaps not?) simplification worked great after all that notation. We
still may want to revisit the proof to use some let-bindings to
simplify the proof steps.